moved IP function, added more docs, fixed coerceToNumberType
