+++ packages/utils/src/ip.ts
@@ -3,7 +3,7 @@ import ipaddr, { IPv4, IPv6 } from 'ipaddr.js';
 import { parse, stringify } from 'ip-bigint';
 import ip6addr from 'ip6addr';
 import validateCidr from 'is-cidr';
-import { isString } from './strings';
+import { isString, primitiveToString } from './strings';
 import { toInteger, isNumberLike, toBigIntOrThrow } from './numbers';
 import { getTypeOf } from './deps';
 
@@ -11,12 +11,24 @@ export function isIP(input: unknown): input is string {
     return isString(input) && _isIP(input);
 }
 
+/** Will throw if input is not a valid CIDR */
 export function isIPRangeOrThrow(input: unknown): string {
     if (!isCIDR(input)) {
         throw new TypeError(`Expected ${input} (${getTypeOf(input)}) to be a valid IP range`);
     }
 
     return input;
+}
+
+/** Will throw if input is not a valid IP */
+export function isIPOrThrow(input: unknown): string {
+    const ipValue = primitiveToString(input);
+
+    if (!isString(input) || !isIP(ipValue)) {
+        throw new TypeError(`Expected ${ipValue} (${getTypeOf(input)}) to be a valid IP`);
+    }
+
+    return ipValue;
 }
 
 export function isIPV6(input: unknown): boolean {
